backend/pyproject.toml
[project]
name = "steam-matchmaker-backend"
version = "1.0.0"
description = "AI-powered game recommendations for Steam friend groups"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "pydantic-ai>=0.0.15",
    "python-dotenv>=1.0.1",
    "groq>=0.11.0",
]

[project.scripts]
start = "uvicorn main:app --reload --port 8000"

[tool.uv]
dev-dependencies = []
